Code modularization and ES6 syntax switch
