+++ .gitignore
@@ -2,3 +2,4 @@
 .hlx/
 logs/
 node_modules/
+npm-debug.log
