+++ fstab.yaml
@@ -1,2 +1,2 @@
 mountpoints:
-  /: https://adobe.sharepoint.com/sites/testnarcis/Shared%20Documents/documents
+  /: https://adobe.sharepoint.com/sites/FedsPublishing/Shared%20Documents/fedPub
