+++ head.html
@@ -1,10 +1,6 @@
-<script>
-    console.log('test log from head.html');
-</script>
-
 <!-- THIS IS THE DEFAULT FILE CONTENT -->
 <link rel="stylesheet" href="/styles.css"/>
 <link id="favicon" rel="icon" type="image/svg+xml" href="/icons/adobe.svg">
 <link rel="stylesheet" href="/hlx_fonts/pnv6nym.css"/>
 <meta name="viewport" content="width=device-width, initial-scale=1"/>
-<script src="/scripts.js" defer></script>
+<script type="module" src="/scripts.js"></script>
