+++ modules/CONFIG.js
@@ -1,0 +1,8 @@
+export const CONFIG = {
+    SELECTORS: {
+        NAMESPACE: 'fedpub',
+        MAIN: 'main',
+        READY: 'ready',
+        SECTION: 'section-wrapper',
+    },
+};
