+++ modules/conversionExamples/convertTables.html
@@ -1,0 +1,38 @@
+<!-- Original HTML -->
+<table>
+    <thead>
+        <tr>
+            <th>Component name</th>
+        </tr>
+    </thead>
+    <tbody>
+        <tr>
+            <td>
+                <h1>Headline</h1>
+                <p>Paragraph text</p>
+            </td>
+        </tr>
+        <tr>
+            <td>
+                <h2>Subheadline</h2>
+                <u>Underlined text</u>
+            </td>
+        </tr>
+    </tbody>
+</table>
+
+<!-- Converted HTML -->
+<div class="fedpub--component-name">
+    <div class="fedpub--component-name-row">
+        <div class="fedpub--component-name-entry">
+            <h1>Headline</h1>
+            <p>Paragraph text</p>
+        </div>
+    </div>
+    <div class="fedpub--component-name-row">
+        <div class="fedpub--component-name-entry">
+            <h2>Subheadline</h2>
+            <u>Underlined text</u>
+        </div>
+    </div>
+</div>
