--- outline.js
@@ -1,1 +1,0 @@
-console.log('test from outline.js script');
